Fix bug with extract_ttlEphys_timestamps
The function extract_ttlEphys_timestamps was defined but never called. It also contained an error: variable Fs used without being defined.

audio_rec.ttl_ephys.downsampling was set to 100, but the actual downsampling used was 50. Now fixed to 100.
The value of audio_rec.ttl_ephys.SampleRate that was saved is correct.